breakup clients file
