+++ sdk/tables/azure-data-tables/CMakeLists.txt
@@ -50,13 +50,15 @@ set(
     inc/azure/data/tables/internal/policies/shared_key_lite_policy.hpp
     inc/azure/data/tables/internal/policies/tenant_bearer_token_policy.hpp
     inc/azure/data/tables/internal/policies/timeout_policy.hpp
+    inc/azure/data/tables/internal/constants.hpp
     inc/azure/data/tables/internal/serializers.hpp
     inc/azure/data/tables/internal/xml_wrapper.hpp
     inc/azure/data/tables/models.hpp
     inc/azure/data/tables/rtti.hpp
     inc/azure/data/tables.hpp
-    inc/azure/data/tables/tables_clients.hpp
+    inc/azure/data/tables/tables_client.hpp
     inc/azure/data/tables/tables_sas_builder.hpp
+    inc/azure/data/tables/tables_service_client.hpp
 )
 
 set(
