+++ sdk/tables/azure-data-tables/test/ut/table_client_test.hpp
@@ -3,7 +3,8 @@
 // Licensed under the MIT License.
 
 #include "azure/data/tables/models.hpp"
-#include "azure/data/tables/tables_clients.hpp"
+#include "azure/data/tables/tables_client.hpp"
+#include "azure/data/tables/tables_service_client.hpp"
 #include "test/ut/test_base.hpp"
 
 namespace Azure { namespace Data { namespace Test {
